Non visual for DepthStencil, IndexBuffer and OcclusionQuery
