+++ Test/Framework/Graphics/DepthStencilStateNonVisualTest.cs
@@ -1,0 +1,62 @@
+﻿using System;
+using Microsoft.Xna.Framework.Graphics;
+using NUnit.Framework;
+
+namespace MonoGame.Tests.Graphics
+{
+    [TestFixture]
+    internal class DepthStencilStateNonVisualTest : GraphicsDeviceTestFixtureBase
+    {
+        [Test]
+        public void ShouldNotBeAbleToSetNullDepthStencilState()
+        {
+            Assert.Throws<ArgumentNullException>(() => gd.DepthStencilState = null);
+        }
+
+        [Test]
+        public void ShouldNotBeAbleToMutateStateObjectAfterBindingToGraphicsDevice()
+        {
+            var depthStencilState = new DepthStencilState();
+
+            // Can mutate before binding.
+            DoAsserts(depthStencilState, Assert.DoesNotThrow);
+
+            // Can't mutate after binding.
+            gd.DepthStencilState = depthStencilState;
+            DoAsserts(depthStencilState, d => Assert.Throws<InvalidOperationException>(d));
+
+            // Even after changing to different RasterizerState, you still can't mutate a previously-bound object.
+            gd.DepthStencilState = DepthStencilState.Default;
+            DoAsserts(depthStencilState, d => Assert.Throws<InvalidOperationException>(d));
+        }
+
+        [Test]
+        public void ShouldNotBeAbleToMutateDefaultStateObjects()
+        {
+            DoAsserts(DepthStencilState.Default, d => Assert.Throws<InvalidOperationException>(d));
+            DoAsserts(DepthStencilState.DepthRead, d => Assert.Throws<InvalidOperationException>(d));
+            DoAsserts(DepthStencilState.None, d => Assert.Throws<InvalidOperationException>(d));
+        }
+
+        private static void DoAsserts(DepthStencilState depthStencilState, Action<TestDelegate> assertMethod)
+        {
+            assertMethod(() => depthStencilState.CounterClockwiseStencilDepthBufferFail = StencilOperation.Decrement);
+            assertMethod(() => depthStencilState.CounterClockwiseStencilFail = StencilOperation.Decrement);
+            assertMethod(() => depthStencilState.CounterClockwiseStencilFunction = CompareFunction.Always);
+            assertMethod(() => depthStencilState.CounterClockwiseStencilPass = StencilOperation.Decrement);
+            assertMethod(() => depthStencilState.DepthBufferEnable = true);
+            assertMethod(() => depthStencilState.DepthBufferFunction = CompareFunction.Always);
+            assertMethod(() => depthStencilState.DepthBufferWriteEnable = true);
+            assertMethod(() => depthStencilState.ReferenceStencil = 1);
+            assertMethod(() => depthStencilState.StencilDepthBufferFail = StencilOperation.Decrement);
+            assertMethod(() => depthStencilState.StencilEnable = true);
+            assertMethod(() => depthStencilState.StencilFail = StencilOperation.Decrement);
+            assertMethod(() => depthStencilState.StencilFunction = CompareFunction.Always);
+            assertMethod(() => depthStencilState.StencilMask = 1);
+            assertMethod(() => depthStencilState.StencilPass = StencilOperation.Decrement);
+            assertMethod(() => depthStencilState.StencilWriteMask = 1);
+            assertMethod(() => depthStencilState.TwoSidedStencilMode = true);
+        }
+
+    }
+}
