+++ Test/Framework/Visual/DepthStencilStateTest.cs
@@ -13,71 +13,6 @@ namespace MonoGame.Tests.Visual
     [TestFixture]
     internal class DepthStencilStateTest : VisualTestFixtureBase
     {
-        [Test]
-        public void ShouldNotBeAbleToSetNullDepthStencilState()
-        {
-            var game = new TestGameBase();
-            new GraphicsDeviceManager(game);
-
-            game.InitializeOnly();
-
-            Assert.Throws<ArgumentNullException>(() => game.GraphicsDevice.DepthStencilState = null);
-
-            game.Dispose();
-        }
-
-        [Test]
-        public void ShouldNotBeAbleToMutateStateObjectAfterBindingToGraphicsDevice()
-        {
-            var game = new TestGameBase();
-            new GraphicsDeviceManager(game);
-
-            game.InitializeOnly();
-
-            var depthStencilState = new DepthStencilState();
-
-            // Can mutate before binding.
-            DoAsserts(depthStencilState, Assert.DoesNotThrow);
-
-            // Can't mutate after binding.
-            game.GraphicsDevice.DepthStencilState = depthStencilState;
-            DoAsserts(depthStencilState, d => Assert.Throws<InvalidOperationException>(d));
-
-            // Even after changing to different RasterizerState, you still can't mutate a previously-bound object.
-            game.GraphicsDevice.DepthStencilState = DepthStencilState.Default;
-            DoAsserts(depthStencilState, d => Assert.Throws<InvalidOperationException>(d));
-
-            game.Dispose();
-        }
-
-        [Test]
-        public void ShouldNotBeAbleToMutateDefaultStateObjects()
-        {
-            DoAsserts(DepthStencilState.Default, d => Assert.Throws<InvalidOperationException>(d));
-            DoAsserts(DepthStencilState.DepthRead, d => Assert.Throws<InvalidOperationException>(d));
-            DoAsserts(DepthStencilState.None, d => Assert.Throws<InvalidOperationException>(d));
-        }
-
-        private static void DoAsserts(DepthStencilState depthStencilState, Action<TestDelegate> assertMethod)
-        {
-            assertMethod(() => depthStencilState.CounterClockwiseStencilDepthBufferFail = StencilOperation.Decrement);
-            assertMethod(() => depthStencilState.CounterClockwiseStencilFail = StencilOperation.Decrement);
-            assertMethod(() => depthStencilState.CounterClockwiseStencilFunction = CompareFunction.Always);
-            assertMethod(() => depthStencilState.CounterClockwiseStencilPass = StencilOperation.Decrement);
-            assertMethod(() => depthStencilState.DepthBufferEnable = true);
-            assertMethod(() => depthStencilState.DepthBufferFunction = CompareFunction.Always);
-            assertMethod(() => depthStencilState.DepthBufferWriteEnable = true);
-            assertMethod(() => depthStencilState.ReferenceStencil = 1);
-            assertMethod(() => depthStencilState.StencilDepthBufferFail = StencilOperation.Decrement);
-            assertMethod(() => depthStencilState.StencilEnable = true);
-            assertMethod(() => depthStencilState.StencilFail = StencilOperation.Decrement);
-            assertMethod(() => depthStencilState.StencilFunction = CompareFunction.Always);
-            assertMethod(() => depthStencilState.StencilMask = 1);
-            assertMethod(() => depthStencilState.StencilPass = StencilOperation.Decrement);
-            assertMethod(() => depthStencilState.StencilWriteMask = 1);
-            assertMethod(() => depthStencilState.TwoSidedStencilMode = true);
-        }
-
         [TestCase(false)]
         [TestCase(true)]
         public void VisualTestDepthBufferEnable(bool depthBufferEnable)
